rag_service/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
numpy
httpx
pyjwt
python-dotenv
faiss-cpu
pymupdf
